Fix truncate impl in JS bindings

+++ core/storage/database.rs
@@ -19,7 +19,7 @@ pub trait DatabaseStorage: Send + Sync {
     ) -> Result<()>;
     fn sync(&self, c: Completion) -> Result<()>;
     fn size(&self) -> Result<u64>;
-    fn truncate(&self, len: u64, c: Completion) -> Result<()>;
+    fn truncate(&self, len: usize, c: Completion) -> Result<()>;
 }
 
 #[cfg(feature = "fs")]
@@ -76,7 +76,7 @@ impl DatabaseStorage for DatabaseFile {
     }
 
     #[instrument(skip_all, level = Level::INFO)]
-    fn truncate(&self, len: u64, c: Completion) -> Result<()> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         self.file.truncate(len, c)?;
         Ok(())
     }
@@ -141,7 +141,7 @@ impl DatabaseStorage for FileMemoryStorage {
     }
 
     #[instrument(skip_all, level = Level::INFO)]
-    fn truncate(&self, len: u64, c: Completion) -> Result<()> {
+    fn truncate(&self, len: usize, c: Completion) -> Result<()> {
         let _ = self.file.truncate(len, c)?;
         Ok(())
     }
